examples/rag-openai-python/requirements.txt
blindfold-sdk>=1.3.0
openai>=1.0.0
chromadb>=0.5.0
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0
